Refactor dnd client
