+++ clients/zdnd/dnd.h
@@ -1,5 +1,5 @@
-#ifndef ZEN_CLIENT_SIMPLE_BOX_BOX_H
-#define ZEN_CLIENT_SIMPLE_BOX_BOX_H
+#ifndef ZEN_CLIENT_ZDND_H
+#define ZEN_CLIENT_ZDND_H
 
 #include <zukou.h>
 
@@ -7,15 +7,20 @@
 #include <string>
 #include <vector>
 
-#include "stl-object.h"
 #include "types.h"
 
-class Box : public zukou::CuboidWindow
+#pragma pack(1)
+struct StlTriangle {
+  float n[3];
+  float points[3][3];
+  uint16_t unused;
+};
+#pragma pack()
+
+class ZDnd : public zukou::CuboidWindow
 {
  public:
-  Box(zukou::App *app, std::string filename, float length);
-  void Frame(uint32_t time);
-  void Show(std::vector<StlTriangle> triangles);
+  ZDnd(zukou::App *app, std::string filename, float length);
   virtual void DataOfferOffer(const char *mime_type);
   virtual void DataOfferSourceActions(uint32_t source_actions);
   virtual void DataOfferAction(uint32_t dnd_action);
@@ -30,20 +35,31 @@ class Box : public zukou::CuboidWindow
   virtual void RayMotion(uint32_t time, glm::vec3 origin, glm::vec3 direction);
   virtual void RayButton(uint32_t serial, uint32_t time, uint32_t button,
       enum zgn_ray_button_state state);
-  void DrawTexture();
-  void Configure(uint32_t serial, glm::vec3 half_size);
 
-  struct wl_array mime_type_list;
+  bool MainLoop();
+  void Configure(uint32_t serial, glm::vec3 half_size);
+  void InitGL();
+  void UpdateVertex();
+
+  void (*epoll_func)(ZDnd *zdnd);
+  int epoll_fd;
+  int dnd_fd;
+
+  uint32_t drag_enter_serial;
+
   struct zgn_data_source *data_source;
+
+  std::vector<StlTriangle> triangle_list;
+  std::vector<std::string> mime_type_list;
 
  private:
   const std::string path_;
 
+  bool running_;
+
   zukou::OpenGLComponent *component_;
   zukou::OpenGLVertexBuffer *vertex_buffer_;
   zukou::OpenGLShaderProgram *shader_;
-
-  uint32_t triangle_count_;
 
   std::vector<Vertex> vertices_;
 
@@ -53,13 +69,11 @@ class Box : public zukou::CuboidWindow
   } ray_;
   bool ray_focus_;
 
-  uint32_t drag_enter_serial_;
-
   glm::vec3 min_;
   glm::vec3 max_;
 };
 
-#endif  //  ZEN_CLIENT_SIMPLE_BOX_BOX_H
+#endif  //  ZEN_CLIENT_ZDND_BOX_H
 
 // Box.points_
 //                                         /
