+++ clients/zdnd/file-reader.h
@@ -1,5 +1,5 @@
-#ifndef ZEN_CLIENT_ZPREVIEW_FILE_READER_H
-#define ZEN_CLIENT_ZPREVIEW_FILE_READER_H
+#ifndef ZEN_CLIENT_ZDND_FILE_READER_H
+#define ZEN_CLIENT_ZDND_FILE_READER_H
 
 #include <fstream>
 #include <string>
@@ -19,4 +19,4 @@ class FileReader : public Reader
   std::ifstream ifs_;
 };
 
-#endif  //  ZEN_CLIENT_ZPREVIEW_FILE_READER_H
+#endif  //  ZEN_CLIENT_ZDND_FILE_READER_H
