+++ clients/zdnd/viewer.h
@@ -1,5 +1,5 @@
-#ifndef ZEN_CLIENT_ZPREVIEW_VIEWER_H
-#define ZEN_CLIENT_ZPREVIEW_VIEWER_H
+#ifndef ZEN_CLIENT_ZDND_VIEWER_H
+#define ZEN_CLIENT_ZDND_VIEWER_H
 
 class ViewerInterface
 {
@@ -8,4 +8,4 @@ class ViewerInterface
   virtual bool Show() { return true; }
 };
 
-#endif  //  ZEN_CLIENT_ZPREVIEW_VIEWER_H
+#endif  //  ZEN_CLIENT_ZDND_VIEWER_H
